backend/requirements.txt
fastapi==0.104.1
google-generativeai==0.4.1
qdrant-client==1.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
uvicorn==0.24.0
pytest==7.4.3
httpx==0.25.2